+++ src/ui_basic/button.cc
@@ -161,7 +161,7 @@ void Button::draw(RenderTarget& dst) {
 
 	// Draw the background
 	if (pic_background_) {
-		dst.fill_rect(Rectf(0.f, 0.f, get_w(), get_h()), RGBAColor(0, 0, 0, 255));
+		dst.fill_rect(Recti(0, 0, get_w(), get_h()), RGBAColor(0, 0, 0, 255));
 		dst.tile(
 		   Recti(Vector2i(0, 0), get_w(), get_h()), pic_background_, Vector2i(get_x(), get_y()));
 	}
@@ -233,32 +233,32 @@ void Button::draw(RenderTarget& dst) {
 			//  left edge
 			dst.brighten_rect(Recti(0, 2, 2, get_h() - 2), BUTTON_EDGE_BRIGHT_FACTOR);
 			//  bottom edge
-			dst.fill_rect(Rectf(2, get_h() - 2, get_w() - 2, 1), black);
-			dst.fill_rect(Rectf(1, get_h() - 1, get_w() - 1, 1), black);
+			dst.fill_rect(Recti(2, get_h() - 2, get_w() - 2, 1), black);
+			dst.fill_rect(Recti(1, get_h() - 1, get_w() - 1, 1), black);
 			//  right edge
-			dst.fill_rect(Rectf(get_w() - 2, 2, 1, get_h() - 2), black);
-			dst.fill_rect(Rectf(get_w() - 1, 1, 1, get_h() - 1), black);
+			dst.fill_rect(Recti(get_w() - 2, 2, 1, get_h() - 2), black);
+			dst.fill_rect(Recti(get_w() - 1, 1, 1, get_h() - 1), black);
 		} else {
 			//  bottom edge
 			dst.brighten_rect(Recti(0, get_h() - 2, get_w(), 2), BUTTON_EDGE_BRIGHT_FACTOR);
 			//  right edge
 			dst.brighten_rect(Recti(get_w() - 2, 0, 2, get_h() - 2), BUTTON_EDGE_BRIGHT_FACTOR);
 			//  top edge
-			dst.fill_rect(Rectf(0, 0, get_w() - 1, 1), black);
-			dst.fill_rect(Rectf(0, 1, get_w() - 2, 1), black);
+			dst.fill_rect(Recti(0, 0, get_w() - 1, 1), black);
+			dst.fill_rect(Recti(0, 1, get_w() - 2, 1), black);
 			//  left edge
-			dst.fill_rect(Rectf(0, 0, 1, get_h() - 1), black);
-			dst.fill_rect(Rectf(1, 0, 1, get_h() - 2), black);
+			dst.fill_rect(Recti(0, 0, 1, get_h() - 1), black);
+			dst.fill_rect(Recti(1, 0, 1, get_h() - 2), black);
 		}
 	} else {
 		//  Button is flat, do not draw borders, instead, if it is pressed, draw
 		//  a box around it.
 		if (enabled_ && highlighted_) {
 			RGBAColor shade(100, 100, 100, 80);
-			dst.fill_rect(Rectf(0, 0, get_w(), 2), shade);
-			dst.fill_rect(Rectf(0, 2, 2, get_h() - 2), shade);
-			dst.fill_rect(Rectf(0, get_h() - 2, get_w(), get_h()), shade);
-			dst.fill_rect(Rectf(get_w() - 2, 0, get_w(), get_h()), shade);
+			dst.fill_rect(Recti(0, 0, get_w(), 2), shade);
+			dst.fill_rect(Recti(0, 2, 2, get_h() - 2), shade);
+			dst.fill_rect(Recti(0, get_h() - 2, get_w(), get_h()), shade);
+			dst.fill_rect(Recti(get_w() - 2, 0, get_w(), get_h()), shade);
 		}
 	}
 }
