+++ src/ui_basic/panel.cc
@@ -1076,8 +1076,8 @@ bool Panel::draw_tooltip(RenderTarget& dst, const std::string& text) {
 	if (screen_bottom_right.y < tooltip_bottom_right.y)
 		r.y -= 35 + r.h;
 
-	dst.fill_rect(r.cast<float>(), RGBColor(63, 52, 34));
-	dst.draw_rect(r.cast<float>(), RGBColor(0, 0, 0));
+	dst.fill_rect(r, RGBColor(63, 52, 34));
+	dst.draw_rect(r, RGBColor(0, 0, 0));
 	dst.blit(r.origin() + Vector2i(2, 2), rendered_text);
 	return true;
 }
